Split mongoose into two connections (core+instance)
