+++ src/connections/mongoose/index.js
@@ -1,0 +1,4 @@
+const core = require('./core');
+const instance = require('./instance');
+
+module.exports = { core, instance };
