+++ src/models/advertiser.js
@@ -1,4 +1,4 @@
-const mongoose = require('../mongoose');
+const mongoose = require('../connections/mongoose/instance');
 const schema = require('../schema/advertiser');
 
 module.exports = mongoose.model('advertiser', schema);
