+++ src/models/analytics/bot.js
@@ -1,4 +1,4 @@
-const mongoose = require('../../mongoose');
+const mongoose = require('../../connections/mongoose/instance');
 const schema = require('../../schema/analytics/bot');
 
 module.exports = mongoose.model('analytics-bot', schema);
