+++ src/models/analytics/click.js
@@ -1,4 +1,4 @@
-const mongoose = require('../../mongoose');
+const mongoose = require('../../connections/mongoose/instance');
 const schema = require('../../schema/analytics/click');
 
 module.exports = mongoose.model('analytics-click', schema);
