+++ src/models/analytics/event.js
@@ -1,4 +1,4 @@
-const mongoose = require('../../mongoose');
+const mongoose = require('../../connections/mongoose/instance');
 const schema = require('../../schema/analytics/event');
 
 module.exports = mongoose.model('analytics-event', schema);
