+++ src/models/analytics/load.js
@@ -1,4 +1,4 @@
-const mongoose = require('../../mongoose');
+const mongoose = require('../../connections/mongoose/instance');
 const schema = require('../../schema/analytics/load');
 
 module.exports = mongoose.model('analytics-load', schema);
