+++ src/models/analytics/request-object.js
@@ -1,4 +1,4 @@
-const mongoose = require('../../mongoose');
+const mongoose = require('../../connections/mongoose/instance');
 const schema = require('../../schema/analytics/request-object');
 
 module.exports = mongoose.model('analytics-request-object', schema);
