+++ src/models/analytics/request.js
@@ -1,4 +1,4 @@
-const mongoose = require('../../mongoose');
+const mongoose = require('../../connections/mongoose/instance');
 const schema = require('../../schema/analytics/request');
 
 module.exports = mongoose.model('analytics-request', schema);
