+++ src/models/analytics/view.js
@@ -1,4 +1,4 @@
-const mongoose = require('../../mongoose');
+const mongoose = require('../../connections/mongoose/instance');
 const schema = require('../../schema/analytics/view');
 
 module.exports = mongoose.model('analytics-view', schema);
