+++ src/models/campaign.js
@@ -1,4 +1,4 @@
-const mongoose = require('../mongoose');
+const mongoose = require('../connections/mongoose/instance');
 const schema = require('../schema/campaign');
 
 module.exports = mongoose.model('campaign', schema);
