+++ src/models/contact.js
@@ -1,4 +1,4 @@
-const mongoose = require('../mongoose');
+const mongoose = require('../connections/mongoose/instance');
 const schema = require('../schema/contact');
 
 module.exports = mongoose.model('contact', schema);
