+++ src/models/image.js
@@ -1,4 +1,4 @@
-const mongoose = require('../mongoose');
+const mongoose = require('../connections/mongoose/instance');
 const schema = require('../schema/image');
 
 module.exports = mongoose.model('image', schema);
