+++ src/models/placement.js
@@ -1,4 +1,4 @@
-const mongoose = require('../mongoose');
+const mongoose = require('../connections/mongoose/instance');
 const schema = require('../schema/placement');
 
 module.exports = mongoose.model('placement', schema);
