+++ src/models/publisher.js
@@ -1,4 +1,4 @@
-const mongoose = require('../mongoose');
+const mongoose = require('../connections/mongoose/instance');
 const schema = require('../schema/publisher');
 
 module.exports = mongoose.model('publisher', schema);
