+++ src/models/story.js
@@ -1,4 +1,4 @@
-const mongoose = require('../mongoose');
+const mongoose = require('../connections/mongoose/instance');
 const schema = require('../schema/story');
 
 module.exports = mongoose.model('story', schema);
