+++ src/models/template.js
@@ -1,4 +1,4 @@
-const mongoose = require('../mongoose');
+const mongoose = require('../connections/mongoose/instance');
 const schema = require('../schema/template');
 
 module.exports = mongoose.model('template', schema);
