+++ src/models/user.js
@@ -1,4 +1,4 @@
-const mongoose = require('../mongoose');
+const mongoose = require('../connections/mongoose/instance');
 const schema = require('../schema/user');
 
 module.exports = mongoose.model('user', schema);
