+++ src/schema/placement.js
@@ -1,5 +1,5 @@
 const { Schema } = require('mongoose');
-const connection = require('../mongoose');
+const connection = require('../connections/mongoose/instance');
 const { applyElasticPlugin, setEntityFields } = require('../elastic/mongoose');
 
 const schema = new Schema({
