+++ src/schema/story.js
@@ -1,6 +1,6 @@
 const { Schema } = require('mongoose');
 const slug = require('slug');
-const connection = require('../mongoose');
+const connection = require('../connections/mongoose/instance');
 const { applyElasticPlugin, setEntityFields } = require('../elastic/mongoose');
 const imagePlugin = require('../plugins/image');
